Iterator version of rk solvers
