--- src/runge_kutta.jl
@@ -1,16 +1,0 @@
-# Explicit Runge-Kutta solvers
-##############################
-# (Hairer & Wanner 1992 p.134, p.165-169)
-
-import Base: start, next, done, call
-
-include("tableaus.jl")
-
-# include("algorithms.jl")
-
-include("iterators.jl")
-
-# include("iterators.jl")
-# include("fixed.jl")
-# include("variable.jl")
-include("dense.jl")
